Implement a way to display, preview and delete filters
Previously it was not possible to display an users
created filters, and no way to delete them either.

This patch implements this functionality as well as the
possibility to preview ("try-out") created filters the
same way that you can when "building" a new filter.

It also refactors a lot to make it more future-proof,
and fixes some bugs, such as the one where useEffect
is misused resulting in reloading alerts on each
character changed in the filter name...