Remove all linter errors (by fixing or ignoring)
Due to the previous total lack of TypeScript typing
there is still a lot of of code without proper use of
typing. By ignoring the places where we know that there
are a lack of typing we can use the linter for useful
things.

The plan is to gradually introduce typing to the rest of
the code, and to only then enable the linter again.

To find the places where linter errors are ignored use the
following (or search for "eslint-disable") :

```sh
grep -r eslint-disable src
```