Switch from "alert" to "incident" and changes to match backend
Rename all uses of the term "alert" with "incident" and make
required changes to match with latest changes in the backend.

This is a disaster of a commit. I both do a lot of changes
to alerttable (now incidenttable) and renames and other things.

A good example of how not do use git. I'm sorry :(