Fix mistake in ticket_url API usage
The PATCH ticket_url API endpoint does not
return the entire incident, as the api method
in the api client specified. Instead it only
returns the ticket_url itself.

This resulted in the frontend not having
updating the ticket_url when set. This
resolves this by merging the old incident
dict with the new ticket_url response dict.