Fix and improve filters view
This patch introduces fixes to the filter view. It
updates it to work correctly with the changes in the API
and also adds functionality for editing existing filters.

All in all this patch seems to work quite well, with
snackbar alerts used to provide the user with feedback.

Doesn't look that good, but does the job.