Fix disappearance of open-incident confirm dialog after interval update
Reopen-incident confirmation dialog (as well as any other dialog related to incident modification) disappears only if same action (same incident update) was performed by some other user in a meanwhile. Otherwise any already opened dialog remains on the page even if incidents table is updated. 
The fix introduced in this commit is making ConfirmationButton both a stateful component (used for open-incident dialog), as well as a functional component (used for any of the other dialogs).