Add cookie domain to environment vars and fetch it in config
Providing cookie domain in config is necessary for proper removal of dataporten-token. Dataporten-token is stored as subdomain and can not be removed unless frontend knows the domain name. This could be extracted from the url, but since cookie domain is provided in api settings (which theoretically can be set to a value that differs from the api url), frontend should not assume that cookie domain can be extracted from the url.