Default cookie domain to backend's hostname
COOKIE_DOMAIN on frontend will be set to one of the following (first one that is a truthy value): REACT_APP_COOKIE_DOMAIN if it is provided, or hostname extracted from REACT_APP_BACKEND_URL if latter is provided, or hostname extracted from url frontend is running on. The last scenario is a backup and should ideally not occur, since backend url de facto must be provided.