Make SVG logos render correctly
The SVG logos contained text objects using a font that isn't normally
available on all systems. This caused the logos to be rendered
incorrectly on many systems, as the text would be rendered with an
alternative font which was wider - causing clipping of the logo type.

This replaced the text objects with paths, so the logos will always
render correctly.

Fixes #366