Rework production docker image for runtime config
This updates the docker production image to be a fully operational nginx
server which serves the statically built Argus front-end.  Now that
runtime configuration is supported, the updated image will produce a
runtime config file from the same environment variables that were
originally expected by the frontend build process, and ensure it's in
the correct location before starting nginx.